ürünler veritabanına eklendi

+++ SupermarketManagementSystem/database/debt/Debt.cs
@@ -8,7 +8,7 @@ namespace SupermarketManagementSystem.database
 {
 	class Debt
 	{
-		public int debt_amount { get; set; }
+		public float debt_amount { get; set; }
 		public bool payed { get; set; } //borç ödendiyse true olacak 
 		public DateTime debt_date { get; set; }
